pyproject.toml
[project]
name = "repops"
version = "0.1.0"
description = "A tool for managing multiple code repositories"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pytest>=7.0.0",
    "flake8>=6.0.0",
    "autopep8>=2.0.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
repops = "repops.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
